refactor react card context
